refactor: moved all internal use-cases inside `application.internal`
