+++ src/main/java/io/pakland/mdas/githubstats/application/internal/AggregateCommits.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.internal;
 
 import io.pakland.mdas.githubstats.domain.entity.Commit;
 import io.pakland.mdas.githubstats.domain.entity.CommitAggregation;
