+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/controller/UserControllerFromDb.java
@@ -1,8 +1,8 @@
 package io.pakland.mdas.githubstats.infrastructure.controller;
 
 import io.pakland.mdas.githubstats.application.FetchAvailableOrganizations;
-import io.pakland.mdas.githubstats.application.GetUserByLogin;
-import io.pakland.mdas.githubstats.application.OrchestrateAggregators;
+import io.pakland.mdas.githubstats.application.internal.GetUserByLogin;
+import io.pakland.mdas.githubstats.application.internal.OrchestrateAggregators;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
 import io.pakland.mdas.githubstats.application.exceptions.UserLoginNotFound;
 import io.pakland.mdas.githubstats.domain.entity.Organization;
