+++ src/test/java/io/pakland/mdas/githubstats/application/internal/AggregateUserReviewsTest.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.internal;
 
 import io.pakland.mdas.githubstats.domain.entity.Comment;
 import io.pakland.mdas.githubstats.domain.entity.UserReview;
