+++ src/test/java/io/pakland/mdas/githubstats/application/internal/SaveAllPullRequestsTest.java
@@ -1,5 +1,6 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.internal;
 
+import io.pakland.mdas.githubstats.application.internal.SaveAllPullRequests;
 import io.pakland.mdas.githubstats.domain.entity.PullRequest;
 import io.pakland.mdas.githubstats.domain.repository.PullRequestRepository;
 import org.junit.jupiter.api.Test;
