+++ src/test/java/io/pakland/mdas/githubstats/application/internal/SaveAllUserReviewsTest.java
@@ -1,5 +1,6 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.internal;
 
+import io.pakland.mdas.githubstats.application.internal.SaveAllUserReviews;
 import io.pakland.mdas.githubstats.domain.entity.UserReview;
 import io.pakland.mdas.githubstats.domain.repository.UserReviewRepository;
 import org.junit.jupiter.api.Test;
