+++ src/test/java/io/pakland/mdas/githubstats/application/internal/SaveAllUsersTest.java
@@ -1,5 +1,6 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.internal;
 
+import io.pakland.mdas.githubstats.application.internal.SaveAllUsers;
 import io.pakland.mdas.githubstats.domain.entity.User;
 import io.pakland.mdas.githubstats.domain.repository.UserRepository;
 import org.junit.jupiter.api.Test;
