+++ src/test/java/io/pakland/mdas/githubstats/application/unused/GetRepositoriesByTeamTest.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.application;
+package io.pakland.mdas.githubstats.application.unused;
 
 import io.pakland.mdas.githubstats.application.unnused.GetRepositoriesByTeam;
 import io.pakland.mdas.githubstats.domain.entity.Repository;
